Moved no copy test cases to own file. Added debugging to cloudwatch helper to triage flakiness
